Remove RemoteChainProcessor calculating noteIndex
